requirements.txt
# OmniParser + Qwen OCR - Portable Requirements
# Works with Python 3.10-3.12, CUDA 12.x
# PyTorch is installed separately by install.py (CUDA version detection)

# === Core ML/Vision ===
transformers==4.49.0
accelerate>=1.2.0
timm>=1.0.0
einops>=0.8.0
ultralytics>=8.3.0
supervision>=0.18.0

# === OCR (PaddleOCR detection + fallback recognition) ===
easyocr>=1.7.0
paddlepaddle>=3.0.0
paddleocr>=2.7.0

# === OpenCV ===
opencv-python>=4.6.0
opencv-python-headless>=4.6.0

# === Web/API Server ===
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
gradio>=5.0.0

# === Utilities ===
numpy>=1.24.0,<2.0
Pillow>=10.0.0
dill>=0.3.0
requests>=2.28.0
safetensors>=0.4.0

# === HuggingFace Hub (for model downloads) ===
huggingface_hub>=0.20.0
